requirements.txt
aniso8601==10.0.1
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Flask==3.1.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pytz==2026.1.post1
six==1.17.0
SQLAlchemy==2.0.48
typing_extensions==4.15.0
Werkzeug==3.1.6
Flask-JWT-Extended
python-dotenv
flask-cors
